feature: money bigdecimal migration (#176)
* feat(merchant acquirer): replace long -> BigDecimal and fix calculate fee

* docs(merchant acquirer): replace long -> BigDecimal

* test: update TransactionRequest DTO and related switch/e2e tests

* refactor(merchant acquirer): edit calculate fee logic and generate amount

* feat(switch): update acquiring fee client, merchant acquirer integration, update route service and logger tests

* style(switch): fix long line in MerchantAcquirerClient for checkstyle

* feat: bigdecimal money type in gateway

* feat(card-management): now using bigdecimal money values

* feat(terminal-simulator): long to BigDecimal

* refactor(transaction-logger): replace Long with BigDecimal for money related fields

* feat: implemented BigDecimal for authorization

* fix(card-management): after-merge issues

* test(switch): add BigDecimal precision test for fractional amount

* fix: removed assignment for LimitUsage.dailyLimit and monthlyLimit

* test: added test for checking long vs BigDecimal

* fix: fixed e2e TC-20, long -> BigDecimal

---------

+++ services/card-management/src/main/java/com/processing/cardmanagement/options/CardGeneratorOptions.java
@@ -3,6 +3,8 @@ package com.processing.cardmanagement.options;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
+import java.math.BigDecimal;
+
 /**
  * Конфигурация генератора тестовых карт
  * Загружается из application.properties с префиксом app.card-generator
@@ -10,9 +12,9 @@ import org.springframework.validation.annotation.Validated;
 @Validated
 @ConfigurationProperties(prefix = "app.card-generator")
 public record CardGeneratorOptions(
-    int minBalance,
-    int maxBalance,
-    int minDailyLimit,
-    int maxDailyLimit,
+    BigDecimal minBalance,
+    BigDecimal maxBalance,
+    BigDecimal minDailyLimit,
+    BigDecimal maxDailyLimit,
     String currencyCode
 ) {}
